refactor: Refactor email, formsubmit forms Link T-12340
